Generator - Extract ScopeInfo from Processor
