+++ inject-generator/src/main/java/io/avaje/inject/generator/ProcessingContext.java
@@ -30,12 +30,6 @@ class ProcessingContext {
   private final Filer filer;
   private final Elements elementUtils;
   private final Types typeUtils;
-
-  private String contextName;
-  private String[] contextProvides;
-  private String[] contextDependsOn;
-  private Set<String> contextRequires = new LinkedHashSet<>();
-  private String contextPackage;
   private String metaInfServicesLine;
 
   ProcessingContext(ProcessingEnvironment processingEnv) {
@@ -112,39 +106,6 @@ class ProcessingContext {
     return filer.createResource(StandardLocation.CLASS_OUTPUT, "", Constants.META_INF_FACTORY);
   }
 
-  void setContextDetails(String name, String[] provides, String[] dependsOn, Element contextElement) {
-    this.contextName = name;
-    this.contextProvides = provides;
-    this.contextDependsOn = dependsOn;
-
-    // determine the context package (that we put the DI Factory class into)
-    PackageElement pkg = elementUtils.getPackageOf(contextElement);
-    logDebug("using package from element " + pkg);
-    this.contextPackage = (pkg == null) ? null : pkg.getQualifiedName().toString();
-  }
-
-  void setContextRequires(List<String> contextRequires) {
-    this.contextRequires.addAll(contextRequires);
-  }
-
-  Set<String> contextRequires() {
-    return contextRequires;
-  }
-
-  void deriveContextName(String factoryPackage) {
-    if (contextName == null) {
-      contextName = factoryPackage;
-    }
-  }
-
-  String contextName() {
-    return contextName;
-  }
-
-  String getContextPackage() {
-    return contextPackage;
-  }
-
   TypeElement element(String rawType) {
     return elementUtils.getTypeElement(rawType);
   }
@@ -161,61 +122,7 @@ class ProcessingContext {
     return typeUtils.asElement(returnType);
   }
 
-  void buildNewBuilder(Append writer) {
-    writer.append("    this.name = \"%s\";", contextName).eol();
-    writer.append("    this.provides = ", contextProvides);
-    buildStringArray(writer, contextProvides, true);
-    writer.append(";").eol();
-    writer.append("    this.dependsOn = ", contextDependsOn);
-    buildStringArray(writer, contextDependsOn, true);
-    writer.append(";").eol();
-  }
-
-  void buildAtInjectModule(Append writer) {
-    writer.append(Constants.AT_GENERATED).eol();
-    writer.append("@InjectModule(name=\"%s\"", contextName);
-    if (!isEmpty(contextProvides)) {
-      writer.append(", provides=");
-      buildStringArray(writer, contextProvides, false);
-    }
-    if (!isEmpty(contextDependsOn)) {
-      writer.append(", dependsOn=");
-      buildStringArray(writer, contextDependsOn, false);
-    }
-    if (!contextRequires.isEmpty()) {
-      writer.append(", requires={");
-      int c = 0;
-      for (String value : contextRequires) {
-        if (c++ > 0) {
-          writer.append(",");
-        }
-        writer.append(value).append(".class");
-      }
-      writer.append("}");
-    }
-    writer.append(")").eol();
-  }
-
-  private boolean isEmpty(String[] strings) {
-    return strings == null || strings.length == 0;
-  }
-
-  private void buildStringArray(Append writer, String[] values, boolean asArray) {
-    if (isEmpty(values)) {
-      writer.append("null");
-    } else {
-      if (asArray) {
-        writer.append("new String[]");
-      }
-      writer.append("{");
-      int c = 0;
-      for (String value : values) {
-        if (c++ > 0) {
-          writer.append(",");
-        }
-        writer.append("\"").append(value).append("\"");
-      }
-      writer.append("}");
-    }
+  PackageElement getPackageOf(Element element) {
+    return elementUtils.getPackageOf(element);
   }
 }
