+++ inject-generator/src/main/java/io/avaje/inject/generator/Processor.java
@@ -1,44 +1,22 @@
 package io.avaje.inject.generator;
 
-import io.avaje.inject.InjectModule;
 import io.avaje.inject.Factory;
-import io.avaje.inject.spi.DependencyMeta;
+import io.avaje.inject.InjectModule;
+import jakarta.inject.Singleton;
 
 import javax.annotation.processing.AbstractProcessor;
-import javax.annotation.processing.FilerException;
 import javax.annotation.processing.ProcessingEnvironment;
 import javax.annotation.processing.RoundEnvironment;
-import jakarta.inject.Singleton;
 import javax.lang.model.SourceVersion;
 import javax.lang.model.element.*;
 import javax.lang.model.util.Elements;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.Iterator;
-import java.util.LinkedHashMap;
-import java.util.LinkedHashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import java.util.*;
 
 public class Processor extends AbstractProcessor {
 
-  private static final String INJECT_MODULE = "io.avaje.inject.InjectModule";
-
   private ProcessingContext context;
-
   private Elements elementUtils;
-
-  /**
-   * Map to merge the existing meta data with partially compiled code. Keyed by type and qualifier/name.
-   */
-  private final Map<String, MetaData> metaData = new LinkedHashMap<>();
-
-  private final List<BeanReader> beanReaders = new ArrayList<>();
-
-  private final Set<String> readBeans = new HashSet<>();
+  private ScopeInfo scopeInfo;
 
   public Processor() {
   }
@@ -53,11 +31,11 @@ public class Processor extends AbstractProcessor {
     super.init(processingEnv);
     this.context = new ProcessingContext(processingEnv);
     this.elementUtils = processingEnv.getElementUtils();
+    this.scopeInfo = new ScopeInfo(context);
   }
 
   @Override
   public Set<String> getSupportedAnnotationTypes() {
-
     Set<String> annotations = new LinkedHashSet<>();
     annotations.add(InjectModule.class.getCanonicalName());
     annotations.add(Factory.class.getCanonicalName());
@@ -83,48 +61,8 @@ public class Processor extends AbstractProcessor {
     readChangedBeans(beans, false);
     readChangedBeans(controllers, false);
 
-    mergeMetaData();
-
-    writeBeanHelpers();
-    if (roundEnv.processingOver()) {
-      writeBeanFactory();
-    }
+    scopeInfo.write(roundEnv.processingOver());
     return false;
-  }
-
-  private void writeBeanHelpers() {
-    for (BeanReader beanReader : beanReaders) {
-      try {
-        if (!beanReader.isWrittenToFile()) {
-          SimpleBeanWriter writer = new SimpleBeanWriter(beanReader, context);
-          writer.write();
-          beanReader.setWrittenToFile();
-        }
-      } catch (FilerException e) {
-        context.logWarn("FilerException to write $DI class " + beanReader.getBeanType() + " " + e.getMessage());
-
-      } catch (IOException e) {
-        e.printStackTrace();
-        context.logError(beanReader.getBeanType(), "Failed to write $DI class");
-      }
-    }
-  }
-
-  private void writeBeanFactory() {
-    MetaDataOrdering ordering = new MetaDataOrdering(metaData.values(), context);
-    int remaining = ordering.processQueue();
-    if (remaining > 0) {
-      ordering.logWarnings();
-    }
-
-    try {
-      SimpleFactoryWriter factoryWriter = new SimpleFactoryWriter(ordering, context);
-      factoryWriter.write();
-    } catch (FilerException e) {
-      context.logWarn("FilerException trying to write factory " + e.getMessage());
-    } catch (IOException e) {
-      context.logError("Failed to write factory " + e.getMessage());
-    }
   }
 
   /**
@@ -135,58 +73,9 @@ public class Processor extends AbstractProcessor {
       if (!(element instanceof TypeElement)) {
         context.logError("unexpected type [" + element + "]");
       } else {
-        if (readBeans.add(element.toString())) {
-          readBeanMeta((TypeElement) element, factory);
-        } else {
-          context.logDebug("skipping already processed bean " + element);
-        }
+        scopeInfo.read((TypeElement) element, factory);
       }
     }
-  }
-
-  /**
-   * Merge the changed bean meta data into the existing (factory) metaData.
-   */
-  private void mergeMetaData() {
-    for (BeanReader beanReader : beanReaders) {
-      if (!beanReader.isRequestScopedController()) {
-        MetaData metaData = this.metaData.get(beanReader.getMetaKey());
-        if (metaData == null) {
-          addMeta(beanReader);
-        } else {
-          updateMeta(metaData, beanReader);
-        }
-      }
-    }
-  }
-
-  /**
-   * Add a new previously unknown bean.
-   */
-  private void addMeta(BeanReader beanReader) {
-    MetaData meta = beanReader.createMeta();
-    metaData.put(meta.getKey(), meta);
-    for (MetaData methodMeta : beanReader.createFactoryMethodMeta()) {
-      metaData.put(methodMeta.getKey(), methodMeta);
-    }
-  }
-
-  /**
-   * Update the meta data on a previously known bean.
-   */
-  private void updateMeta(MetaData metaData, BeanReader beanReader) {
-    metaData.update(beanReader);
-  }
-
-  /**
-   * Read the dependency injection meta data for the given bean.
-   */
-  private void readBeanMeta(TypeElement typeElement, boolean factory) {
-    if (typeElement.getKind() == ElementKind.ANNOTATION_TYPE) {
-      context.logDebug("skipping annotation type " + typeElement);
-      return;
-    }
-    beanReaders.add(new BeanReader(typeElement, context, factory).read());
   }
 
   /**
@@ -208,34 +97,12 @@ public class Processor extends AbstractProcessor {
         Element element = iterator.next();
         InjectModule annotation = element.getAnnotation(InjectModule.class);
         if (annotation != null) {
-          context.setContextDetails(annotation.name(), annotation.provides(), annotation.dependsOn(), element);
-          context.setContextRequires(readRequires(element));
+          scopeInfo.details(annotation.name(), annotation.provides(), annotation.dependsOn(), element);
+          scopeInfo.requires(ScopeUtil.readRequires(element));
         }
       }
     }
   }
-
-  /**
-   * Read the list of required class names.
-   */
-  private List<String> readRequires(Element element) {
-    List<String> requiresList = new ArrayList<>();
-    for (AnnotationMirror annotationMirror : element.getAnnotationMirrors()) {
-      if (INJECT_MODULE.equals(annotationMirror.getAnnotationType().toString())) {
-        for (Map.Entry<? extends ExecutableElement, ? extends AnnotationValue> entry : annotationMirror.getElementValues().entrySet()) {
-          if (entry.getKey().toString().startsWith("requires")) {
-            for (Object requiresType : (List<?>) entry.getValue().getValue()) {
-              String fullName = requiresType.toString();
-              fullName = fullName.substring(0, fullName.length() - 6);
-              requiresList.add(fullName);
-            }
-          }
-        }
-      }
-    }
-    return requiresList;
-  }
-
 
   /**
    * Read the existing factory bean. Each of the build methods is annotated with <code>@DependencyMeta</code>
@@ -243,29 +110,15 @@ public class Processor extends AbstractProcessor {
    */
   private void readFactory(TypeElement factoryType) {
     InjectModule module = factoryType.getAnnotation(InjectModule.class);
-    context.setContextDetails(module.name(), module.provides(), module.dependsOn(), factoryType);
-    context.setContextRequires(readRequires(factoryType));
+    scopeInfo.details(module.name(), module.provides(), module.dependsOn(), factoryType);
+    scopeInfo.requires(ScopeUtil.readRequires(factoryType));
 
     List<? extends Element> elements = factoryType.getEnclosedElements();
     if (elements != null) {
       for (Element element : elements) {
         if (ElementKind.METHOD == element.getKind()) {
-          readBuildMethodDependencyMeta(element);
+          scopeInfo.readBuildMethodDependencyMeta(element);
         }
-      }
-    }
-  }
-
-  private void readBuildMethodDependencyMeta(Element element) {
-    Name simpleName = element.getSimpleName();
-    if (simpleName.toString().startsWith("build_")) {
-      // read a build method - DependencyMeta
-      DependencyMeta meta = element.getAnnotation(DependencyMeta.class);
-      if (meta == null) {
-        context.logError("Missing @DependencyMeta on method " + simpleName);
-      } else {
-        final MetaData metaData = new MetaData(meta);
-        this.metaData.put(metaData.getKey(), metaData);
       }
     }
   }
